Set "eol-style = native" on all .cs files, they had differing linebreaks (will probably mess up blame).
Added a LevelFilteredLogger which is a base logger that has Level and Name properties and redirects all log calls to a Log method which is abstract.

Added LoggerLevel.Off which can be used in LevelFilteredLogger to turn off logging completly.

Refactored ConsoleLogger to use the new LevelFilteredLogger.

Added a StreamLogger which can log to any (writable) stream.

All other changes are non-changes due to the linefeed mismatch in the source files.

git-svn-id: https://svn.castleproject.org/svn/castle/trunk@633 73e77b4c-caa6-f847-a29a-24ab75ae54b6